moved mailer and checkout to index
