+++ package.json
@@ -1,7 +1,8 @@
 {
   "dependencies": {
     "express": "^4.17.1",
-    "sqlite3": "^5.0.2"
+    "sqlite3": "^5.0.2",
+    "nodemailer": "^6.6.5"
   },
   "name": "5V-webshop",
   "version": "1.0.0",
