+++ web/order.html
@@ -57,7 +57,7 @@
                    class="form-control"
                    name="lastName" 
                    placeholder="Achternaam"
-                   oninvalid="this.setCustomValidity('Archternaam is verplicht')" 
+                   oninvalid="this.setCustomValidity('Achternaam is verplicht')" 
                    onchange="this.setCustomValidity('')">
             
           </div>
